requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
pyyaml
watchdog
pytest
pytest-asyncio
networkx
numpy
pandas
scikit-learn
sentence-transformers
redis
celery
requests
aiohttp
tqdm
Pillow
pytesseract
python-dotenv
packaging
psutil
pydantic-settings
anyio
h11
idna
sniffio
starlette
typing-extensions
